Add empirical agent model and refactor UI
Introduce an OD-based empirical agent workflow and modernise the empirical applet/parameters handling.

- Added EmpiricalAgent and EmpiricalGroup to implement OD-trip behaviour on core.Agent and enum labels.
- Rewrote EmpiricalAgentProperties to adapt legacy empirical group parameters to the new AgentProperties API: uses EnumMap, safer probability sampling, clamping/rescaling, weighted sampling helpers and defensive fallbacks.
- Updated EmpiricalAgentsGroup to parse CSV-style calibration rows, expose typed doubles and group share, and use EmpiricalGroup enum labels.
- Added a minimal EmpiricalDisplay panel and new empirical applet entry (PedSimCityEmpiricalApplet) while removing several legacy cityimage applet classes (Display, ParametersPanel, PedSimCityApplet, SpecificODpanel, TestPanel).
- Added EmpiricalImport and EmpiricalPars support files and integrated changes in engine (PedSimCityEmpirical), Populate and Run to hook the empirical workflow.

Purpose: migrate and refactor the legacy empirical ABM code so it works with the core agent/routing APIs, improve robustness of parameter sampling, and replace the old cityimage UI pieces with a compact empirical-friendly display and applet.

+++ src/main/java/pedsim/cityimage/engine/Import.java
@@ -3,64 +3,67 @@ package pedsim.cityimage.engine;
 import java.io.InputStreamReader;
 import java.io.Reader;
 import java.net.URL;
+
 import com.opencsv.CSVReader;
+
 import pedsim.cityimage.parameters.TestPars;
 import pedsim.core.parameters.Pars;
 
 /**
- * This class is responsible for importing various data files required for the simulation based on
- * the selected simulation parameters. It includes methods for importing distances, barriers,
- * landmarks and sight lines, road network graphs, and empirical agent groups data.
+ * This class is responsible for importing various data files required for the
+ * simulation based on the selected simulation parameters. It includes methods
+ * for importing distances, barriers, landmarks and sight lines, road network
+ * graphs, and empirical agent groups data.
  */
 public class Import extends pedsim.core.engine.Import {
 
-  /**
-   * Imports various data files required for the simulation based on the selected simulation
-   * parameters.
-   *
-   * @throws Exception If an error occurs during the import process.
-   */
-  @Override
-  public void importFiles() throws Exception {
+	/**
+	 * Imports various data files required for the simulation based on the selected
+	 * simulation parameters.
+	 *
+	 * @throws Exception If an error occurs during the import process.
+	 */
+	@Override
+	public void importFiles() throws Exception {
 
-    if (TestPars.testingLandmarks) {
-      importDistances();
-      readLandmarksAndSightLines();
-    } else if (TestPars.testingSubdivisions)
-      readBarriers();
-    else if (TestPars.testingModels) {
-      readLandmarksAndSightLines();
-      readBarriers();
-    }
-    // Read the street network shapefiles and create the primal and the dual graph
-    readGraphs();
-  }
+		if (TestPars.testingLandmarks) {
+			importDistances();
+			readLandmarksAndSightLines();
+		} else if (TestPars.testingSubdivisions)
+			readBarriers();
+		else if (TestPars.testingModels) {
+			readLandmarksAndSightLines();
+			readBarriers();
+		}
+		// Read the street network shapefiles and create the primal and the dual graph
+		readGraphs();
+	}
 
-  /**
-   * Imports GPS trajectory-derived distances required for the simulation.
-   *
-   * @throws Exception If an error occurs during the import process.
-   */
-  private void importDistances() throws Exception {
+	/**
+	 * Imports GPS trajectory-derived distances required for the simulation.
+	 *
+	 * @throws Exception If an error occurs during the import process.
+	 */
+	private void importDistances() throws Exception {
 
-    // Read GPS trajectories distances
-    String resourceName = Pars.cityName + "/" + Pars.cityName + "_distances.csv";
-    URL fileUrl = CLASSLOADER.getResource(resourceName);
-    if (fileUrl == null) {
-      throw new IllegalStateException("Resource not found: " + resourceName);
-    }
+		// Read GPS trajectories distances
+		String resourceName = Pars.cityName + "/" + Pars.cityName + "_distances.csv";
+		URL fileUrl = CLASSLOADER.getResource(resourceName);
+		if (fileUrl == null) {
+			throw new IllegalStateException("Resource not found: " + resourceName);
+		}
 
-    Reader reader = new InputStreamReader(fileUrl.openStream());
-    CSVReader readerDistances = new CSVReader(reader);
-    String[] nextLineDistances;
+		Reader reader = new InputStreamReader(fileUrl.openStream());
+		CSVReader readerDistances = new CSVReader(reader);
+		String[] nextLineDistances;
 
-    int row = 0;
-    while ((nextLineDistances = readerDistances.readNext()) != null) {
-      row += 1;
-      if (row == 1)
-        continue; // Skip header
-      PedSimCityImage.distances.add(Float.parseFloat(nextLineDistances[2]));
-    }
-    readerDistances.close();
-  }
+		int row = 0;
+		while ((nextLineDistances = readerDistances.readNext()) != null) {
+			row += 1;
+			if (row == 1)
+				continue; // Skip header
+			PedSimCityImage.distances.add(Double.parseDouble(nextLineDistances[2]));
+		}
+		readerDistances.close();
+	}
 }
